+++ src/Framework/Hooks/CommonConditions.cs
@@ -50,8 +50,30 @@ namespace QuestFramework.Framework.Hooks
                 ["HasMod"] = (valueToCheck, _) => CheckHasModCondition(valueToCheck),
                 ["Random"] = (valueToCheck, _) => Game1.random.NextDouble() < Convert.ToDouble(valueToCheck) / 100, // Chance is in %
                 ["EPU"] = (valueToCheck, _) => CheckEpuCondition(valueToCheck), // For compatibility with EPU conditions
-                ["HasItemInInventory"] = CheckItemInInventoryByTags // Check if player has an item in inventory matches given name or context tags
+                ["HasItemInInventory"] = CheckItemInInventoryByTags, // Check if player has an item in inventory matches given name or context tags
+                ["HasActiveQuest"] = (valueToCheck, _) => HasActiveQuest(valueToCheck?.Split(' ')),
+                ["CurrentLocation"] = (valueToCheck, _) => Game1.player?.currentLocation?.Name == valueToCheck,
             };
+        }
+
+        private static bool HasActiveQuest(string[] vs)
+        {
+            if (vs == null)
+            {
+                return false;
+            }
+
+            foreach (var v in vs)
+            {
+                var quest = QuestFrameworkMod.Instance.QuestManager.Fetch(v);
+
+                if (quest != null && quest.IsInQuestLog())
+                {
+                    return true;
+                }
+            }
+
+            return false;
         }
 
         private static bool CheckItemInInventoryByTags(string valueToCheck, object context)