+++ docs/changelog.md
@@ -1,12 +1,14 @@
 # Changelog
 
 ## Upcoming version
+
 - Added Hungarian translation by martin66789 and spanish translation by Carlos
+- Enhanced logging for loading content packs and exposing quest types
 - Fixed minor issue when resizing while CustomBoard is open
 - Fixed controller input problem on CustomBoards
 - Added better log for used action tiles for custom dropboxes
 - Fixed slay monster quest reward problem
-- Added new common global condition: `HasItemInInventory`
+- Added new common global conditions: `HasItemInInventory`, `HasActiveQuest` and `CurrentLocation`
 - Quest offer manager now internally checks if offering quest is already in log (and offers it ONLY if not)
 - Added offering quest by pick up an item
 - Reworked NPC quest offer system (optimized)